create files

+++ src/main/java/com/desi/SmnApp/entities/City.java
@@ -11,7 +11,7 @@ public class City {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	
 	private String name;
 
@@ -19,17 +19,17 @@ public class City {
 		super();
 	}
 	
-	public City(int id, String name) {
+	public City(Long id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
